requirements.txt
APScheduler
python-telegram-bot
feedparser